Split classes out to reduce loc burder on DrawableCarouselBeatmapSet
